Refactored project structure for configuration files
