+++ src/util/log-entries.js
